Move receipt printer module into utils
